% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mcmc.R
\name{summarize_mcmc}
\alias{summarize_mcmc}
\title{summarize_mcmc}
\usage{
summarize_mcmc(mcmc_object, parameter, g = identity)
}
\arguments{
\item{mcmc_object}{blah}

\item{parameter}{blah}

\item{g}{blah}
}
\value{
A list of estimated variances.
}
\description{
Requires at least 2 datasets. Quantities to input are after transformation by g. All vectors must be ordered by the same datasets.
}
