% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variances.R
\name{estimate_variances}
\alias{estimate_variances}
\title{estimate_variances}
\usage{
estimate_variances(estimates, mcmcses, posterior.variances)
}
\arguments{
\item{estimates}{a vector of estimates for each dataset in the simulation}

\item{mcmcses}{a vector of estimated monte carlo standard errors for each dataset in the simulation}

\item{posterior.variances}{a vector of posterior variances for each dataset in the simulation}
}
\value{
A list of estimated variances.
}
\description{
Requires vectors to be of lenght 2 or more. Quantities to input are after transformation by g. All vectors must be ordered by the same datasets.
}
