% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcse_total.R
\name{mcmcse_total}
\alias{mcmcse_total}
\title{mcmcse_total}
\usage{
mcmcse_total(var_1, var_2, r, m)
}
\arguments{
\item{var_1}{sigma_1^2 in the paper}

\item{var_2}{sigma_2^2 in the paper}

\item{r}{blah}

\item{m}{blah}
}
\value{
A list of optimal r and m.
}
\description{
A function to calculate the total MCMCSE.
}
