% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim}
\alias{optim}
\title{optim}
\usage{
optim(
  time_burnin,
  time_per_es,
  var_1,
  var_2,
  delta = NULL,
  min_m = 1,
  p = NULL
)
}
\arguments{
\item{time_burnin}{average burnin time used by the MCMC}

\item{time_per_es}{average time per effective sample used by the MCMC after burnin}

\item{var_1}{sigma_1^2 in the paper}

\item{var_2}{sigma_2^2 in the paper}

\item{delta}{Delta in the paper}

\item{min_m}{blah}

\item{p}{Delta/sigma_1 in the paper}
}
\value{
A list of optimal r and m.
}
\description{
A function to calculate the optimal allocation by specifying values for all variables.
}
\examples{
optim(1, 1, 0.95*0.05, 1, min_m = 20, p = 0.05)
}
