% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_2.R
\name{optim_2}
\alias{optim_2}
\title{optim_2}
\usage{
optim_2(time_ratio, var_ratio, p, min_m = 1)
}
\arguments{
\item{time_ratio}{average burnin time / average time per effective sample}

\item{var_ratio}{sigma_1^2/sigma_2^2}

\item{p}{Delta/sigma_1 in the paper}

\item{min_m}{blah}
}
\value{
A list of optimal r and m.
}
\description{
A function to calculate the optimal allocation by specifying ratios.
}
\examples{
optim_2(10^5, 1, min_m = 20, p = 0.05)
}
